tests/wpt/mozilla/tests/mozilla/img_width_height_not_rendered.html

<!doctype html>
<meta charset="utf-8">
<title>Image intrinsic dimensions are returned if the image isn't rendered</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="container" style="display: none">
</div>
<script>
async_test(function(t) {
  var img = document.createElement('img');
  img.onload = t.step_func_done(function() {
    assert_equals(img.width, 500, "intrinsic width should've been returned")
    assert_equals(img.height, 378, "intrinsic height should've been returned")
    document.getElementById('container').appendChild(img);
    assert_equals(img.width, 500, "intrinsic width should've been returned");
    assert_equals(img.height, 378, "intrinsic height should've been returned");
  });
  img.src = "test.png";
});
</script>
